finalized login
