+++ frontend/sac-mobile/app/(auth)/welcome.tsx
@@ -4,17 +4,13 @@ import { SafeAreaView, Text, View } from 'react-native';
 import { router } from 'expo-router';
 
 import { Button } from '@/components/button';
-import Wordmark from '@/components/wordmarks';
+import Wordmark from '@/components/wordmark';
 
 const Welcome = () => {
-    const redirect = () => {
-        router.push('/(auth)/login');
-    };
-
     return (
         <SafeAreaView className="flex-col mb-[8%] mx-[8%]">
             <Wordmark />
-            <View className="bg-gray-500 h-[45%] w-ful rounded-xl mt-[5%]"></View>
+            <View className="bg-gray-500 h-[45%] w-ful rounded-xl mt-[5%]" />
             <Text className="h-[18%] text-6xl font-bold mt-[10%]">
                 Welcome to StudCal
             </Text>
