+++ frontend/sac-mobile/app/_layout.tsx
@@ -9,6 +9,7 @@ import FontAwesome from '@expo/vector-icons/FontAwesome';
 
 import { useAuthStore } from '@/hooks/use-auth';
 import { User } from '@/types/user';
+import { View, Text } from 'react-native';
 
 export {
     // Catch any errors thrown by the Layout component.
@@ -17,7 +18,7 @@ export {
 
 export const unstable_settings = {
     // Ensure that reloading on `/modal` keeps a back button present.
-    initialRouteName: '(app)'
+    initialRouteName: ''
 };
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
@@ -41,7 +42,11 @@ export default function RootLayout() {
     }, [loaded]);
 
     if (!loaded) {
-        return null;
+        return (
+            <View>
+                <Text>Loading...</Text>
+            </View>
+        )
     }
 
     return <RootLayoutNav />;
@@ -50,16 +55,25 @@ export default function RootLayout() {
 function RootLayoutNav() {
     const { isLoggedIn, login } = useAuthStore();
 
+
     useEffect(() => {
         const checkLoginStatus = async () => {
-            const accessToken = await getItemAsync('accessToken');
-            const refreshToken = await getItemAsync('refreshToken');
-            const savedUser = await getItemAsync('user');
+            try {
+                const accessToken = await getItemAsync('accessToken');
+                const refreshToken = await getItemAsync('refreshToken');
+                const savedUser = await getItemAsync('user');
 
-            const user: User = savedUser ? JSON.parse(savedUser) : null;
+                console.log('[root] accessToken:', accessToken);
+                console.log('[root] refreshToken:', refreshToken);
 
-            if (accessToken && refreshToken) {
-                login({ accessToken, refreshToken }, user);
+                const user: User = savedUser ? JSON.parse(savedUser) : null;
+
+                if (accessToken && refreshToken) {
+                    // Consider adding token validation (e.g., expiration check)
+                    login({ accessToken, refreshToken }, user);
+                }
+            } catch (error) {
+                console.error('[RootLayoutNav] Error retrieving tokens:', error);
             }
         };
 
@@ -67,14 +81,12 @@ function RootLayoutNav() {
     }, [login]);
 
     useEffect(() => {
-        if (isLoggedIn === null) return;
-
-        if (!isLoggedIn) {
+        if (isLoggedIn === null) {
             router.push('/(auth)/welcome');
             return;
         }
 
-        router.push('/(app)/');
+        router.push(isLoggedIn ? '/(app)/' : '/(auth)/welcome');
     }, [isLoggedIn]);
 
     return (
