+++ frontend/sac-mobile/components/input.tsx
@@ -1,18 +1,11 @@
-import { Text, TextInput, View } from 'react-native';
+import { Text, TextInput, TextInputProps, View } from 'react-native';
 
-interface InputProps {
+interface InputProps extends TextInputProps {
     title: string;
-    placeholder: string;
-    autoCapitalize?: 'sentences' | 'words' | 'characters';
-    autoCorrect: boolean;
-    onChangeText: (...event: any[]) => void;
-    value: string;
-    secureTextEntry?: boolean;
-    onSubmitEditing: () => void;
     error?: boolean;
 }
 
-const Input = ({ title, error, autoCorrect, value, onSubmitEditing, onChangeText, ...props }: InputProps) => {
+const Input = ({ title, error, ...props }: InputProps) => {
     const borderColor = error ? 'border-red-600' : 'border-gray-500';
     return (
         <View>
@@ -20,12 +13,12 @@ const Input = ({ title, error, autoCorrect, value, onSubmitEditing, onChangeText
             <TextInput
                 className={`pt-[4.5%] pb-[4.5%] pl-[5%] w-full border rounded-xl ${borderColor}`}
                 autoCapitalize={props.autoCapitalize || 'none'}
-                autoCorrect={autoCorrect}
+                autoCorrect={props.autoCorrect}
                 placeholder={props.placeholder}
-                onChangeText={onChangeText}
-                value={value}
+                onChangeText={props.onChangeText}
+                value={props.value}
                 secureTextEntry={props.secureTextEntry || false}
-                onSubmitEditing={onSubmitEditing}
+                onSubmitEditing={props.onSubmitEditing}
             />
         </View>
     );
