+++ frontend/sac-mobile/components/wordmark.tsx
@@ -1,0 +1,11 @@
+import { Text, View } from 'react-native';
+
+const Wordmark = () => {
+    return (
+        <View>
+            <Text className={`text-2xl font-bold text-white`}>Wordmark</Text>
+        </View>
+    );
+};
+
+export default Wordmark;
