--- frontend/sac-mobile/components/wordmarks.tsx
@@ -1,29 +1,0 @@
-import { Text, View } from 'react-native';
-
-import { Button } from '@/components/button';
-
-interface WordmarkProps {
-    textColor?: string;
-    backgroundColor?: string;
-    func?: () => void;
-    title?: string;
-}
-
-const Wordmark = (props: WordmarkProps) => {
-    return (
-        <View className="flex flex-row justify-between mx-auto w-full items-center pt-[3%] pb-[5.5%]">
-            <View>
-                <Text className={`text-2xl font-bold ${props.textColor}`}>
-                    Wordmark
-                </Text>
-            </View>
-            {props.func && (
-                <Button onPress={props.func} variant="secondary" size="sm">
-                    {props.title}
-                </Button>
-            )}
-        </View>
-    );
-};
-
-export default Wordmark;
