+++ frontend/sac-mobile/types/item.ts
@@ -1,0 +1,4 @@
+export type Item = {
+    label: string;
+    value: string;
+};
